+++ synth/post_audio/amp.go
@@ -1,6 +1,6 @@
 package post_audio
 
-func Amp(src []float32, val float32) []float32 {
+func Amp(src []float64, val float64) []float64 {
 	for i := range src {
 		src[i] *= val
 	}
